+++ dln.h
@@ -30,8 +30,8 @@
 
 DEPRECATED(char *dln_find_exe(const char*,const char*));
 DEPRECATED(char *dln_find_file(const char*,const char*));
-char *dln_find_exe_r(const char*,const char*,char*,int);
-char *dln_find_file_r(const char*,const char*,char*,int);
+char *dln_find_exe_r(const char*,const char*,char*,size_t);
+char *dln_find_file_r(const char*,const char*,char*,size_t);
 
 #ifdef USE_DLN_A_OUT
 extern char *dln_argv0;
